Refactor settings model and update service imports; add usage controller and metrics service tests
