+++ backend/src/controllers/metrics.controller.ts
@@ -1,33 +1,112 @@
 import { Request, Response } from 'express';
-import { Metrics, Breakdown } from '../models/metrics.model';
+import { MetricDaily, MetricDotcomChatMetrics, MetricDotcomChatModelStats, MetricEditor, MetricIdeChatEditor, MetricIdeChatMetrics, MetricIdeChatModelStats, MetricIdeCompletions, MetricLanguageStats, MetricModelStats, MetricPrMetrics, MetricPrModelStats, MetricPrRepository } from '../models/metrics.model';
+import { Op } from 'sequelize';
 
 class MetricsController {
   // Get all metrics 📊
-  async getAllMetrics(req: Request, res: Response): Promise<void> {
+  async getMetrics(req: Request, res: Response): Promise<void> {
+    const { type, since, until, editor, language, model } = req.query;
     try {
-      const metrics = await Metrics.findAll({
-        include: [Breakdown]
+      // consider the fact that these are UTC dates...
+      const dateFilter: any = { };
+      if (since) {
+        dateFilter[Op.gte] = new Date(since as string);
+      }
+      if (until) {
+        dateFilter[Op.lte] = new Date(until as string);
+      }
+
+      const include = [];
+      const types = type ? (type as string).split(/[ ,]+/) : [];
+      if (types.length === 0 || types.includes('copilot_ide_code_completions')) {
+        include.push({
+          attributes: { exclude: ['id', 'daily_metric_id'] },
+          model: MetricIdeCompletions,
+          as: 'copilot_ide_code_completions',
+          include: [{
+            attributes: { exclude: ['id', 'ide_completion_id'] },
+            model: MetricEditor,
+            as: 'editors',
+            where: editor ? { name: editor } : {},
+            required: false,
+            include: [{
+              attributes: { exclude: ['id', 'editor_id'] },
+              model: MetricModelStats,
+              as: 'models',
+              where: model ? { name: model } : {},
+              required: false,
+              include: [{
+                attributes: { exclude: ['id', 'model_stat_id'] },
+                model: MetricLanguageStats,
+                as: 'languages',
+                where: language ? { name: language } : {},
+                required: false,
+              }]
+            }]
+          }]
+        });
+      }
+      if (types.length === 0 || types.includes('copilot_ide_chat')) {
+        include.push({
+          attributes: { exclude: ['id', 'daily_metric_id'] },
+          model: MetricIdeChatMetrics,
+          as: 'copilot_ide_chat',
+          include: [{
+            attributes: { exclude: ['id', 'chat_metrics_id'] },
+            model: MetricIdeChatEditor,
+            as: 'editors',
+            where: editor ? { name: editor } : {},
+            required: false,
+            include: [{
+              attributes: { exclude: ['id', 'editor_id'] },
+              model: MetricIdeChatModelStats,
+              as: 'models',
+              where: model ? { name: model } : {},
+              required: false,
+            }]
+          }]
+        });
+      }
+      if (types.length === 0 || types.includes('copilot_dotcom_chat')) {
+        include.push({
+          attributes: { exclude: ['id', 'daily_metric_id'] },
+          model: MetricDotcomChatMetrics,
+          as: 'copilot_dotcom_chat',
+          include: [{
+            attributes: { exclude: ['id', 'chat_metrics_id'] },
+            model: MetricDotcomChatModelStats,
+            as: 'models',
+            where: model ? { name: model } : {},
+            required: false,
+          }]
+        });
+      }
+      if (types.length === 0 || types.includes('copilot_dotcom_pull_requests')) {
+        include.push({
+          attributes: { exclude: ['id', 'daily_metric_id'] },
+          model: MetricPrMetrics,
+          as: 'copilot_dotcom_pull_requests',
+          include: [{
+            attributes: { exclude: ['id', 'pr_metrics_id'] },
+            model: MetricPrRepository,
+            as: 'repositories',
+            include: [{
+              attributes: { exclude: ['id', 'repository_id'] },
+              model: MetricPrModelStats,
+              as: 'models',
+              where: model ? { name: model } : {},
+              required: false,
+            }]
+          }]
+        });
+      }
+      const metrics = await MetricDaily.findAll({
+        where: Object.getOwnPropertySymbols(dateFilter).length ? { date: dateFilter } : {},
+        include
       });
       res.status(200).json(metrics); // 🎉 All metrics retrieved!
     } catch (error) {
-      res.status(500).json(error); // 🚨 Error handling
-    }
-  }
-
-  // Get metrics by day 📅
-  async getMetricsByDay(req: Request, res: Response): Promise<void> {
-    try {
-      const { day } = req.params;
-      const metrics = await Metrics.findOne({
-        where: { day },
-        include: [Breakdown]
-      });
-      if (metrics) {
-        res.status(200).json(metrics); // 🎉 Metrics found!
-      } else {
-        res.status(404).json({ error: 'Metrics not found' }); // 🚨 Metrics not found
-      }
-    } catch (error) {
+      console.log(error)
       res.status(500).json(error); // 🚨 Error handling
     }
   }
